put stream into its own file
